refactor: split display module by concept
