+++ src/anonymizer/interface/display/__init__.py
@@ -1,0 +1,10 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Render anonymizer result records for notebook display."""
+
+from __future__ import annotations
+
+from anonymizer.interface.display.record_html import render_record_html
+
+__all__ = ["render_record_html"]
